Update audio type to MediaStream in Message object
Updated the type of the `audio` field in the `Message` object from `Blob` to `MediaStream`. The change has been made for the sake of supporting better audio quality for the received message. Added a `getAudio` function in `MediaTools` which uses `navigator.mediaDevices` to get the audio stream from user's device. This will allow us to get the best possible audio quality from the user's device. In the `app.ts`, added `videoHandler` and `audioHandler` to handle video and audio messages.