Refactor geolocation handling and fix Popup's regex validator
In the changes made, the handling of fetching and displaying geolocation data in `Message.ts` was refactored and moved to its own separate method, named `getGeo`. This was done to decouple the inclusion of geolocation data from other elements of message rendering and to improve code readability and maintainability.

In `MediaTools.ts`, similarly, the handling for stopping of media recording was refactored into its own method, `onDestroy`, to make the code cleaner.

The regex used in `Popup.ts` for longitude and latitude validation was corrected to make it properly validate a string representing geolocation coordinates.